This version of OpenFOAM(+) lost the tmp::operator() that returned a const-reference. As a workaround the code has now been sprinkled with const_cast to make it compatible with this and previous verions of OpenFOAM
WHY the operator() was removed and replaced with a .ref()-method is a
mystery. Supposedly it was because certain compilers were giving
problems. Anyway. It is futile to rant about this

What is clear is that swak won't compile with an unmodified source of
OpenFOAM+ v1606 as
 src/finiteVolume/finiteVolume/fvc/fvcD2dt2.C
 src/finiteVolume/fvMatrices/fvMatrix/fvMatrixSolve.C
are missing some .ref()-calls